fix: PK 모두 Long으로 수정

+++ meerkatai/src/main/java/com/capstone/meerkatai/anomalybehavior/entity/AnomalyBehavior.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 public class AnomalyBehavior {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer anomalyId;
+    private Long anomalyId;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
