+++ meerkatai/src/main/java/com/capstone/meerkatai/storagespace/repository/StorageSpaceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StorageSpaceRepository extends JpaRepository<StorageSpace, Integer> {
+public interface StorageSpaceRepository extends JpaRepository<StorageSpace, Long> {
     Optional<StorageSpace> findByUserUserId(Integer userId);
 }
