+++ meerkatai/src/main/java/com/capstone/meerkatai/streamingvideo/repository/StreamingVideoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface StreamingVideoRepository extends JpaRepository<StreamingVideo, Integer> {
+public interface StreamingVideoRepository extends JpaRepository<StreamingVideo, Long> {
     List<StreamingVideo> findByUserUserId(Integer userId);
     List<StreamingVideo> findByCctvCctvId(Integer cctvId);
 }
