+++ meerkatai/src/main/java/com/capstone/meerkatai/video/dto/VideoDetailsResponse.java
@@ -13,10 +13,10 @@ public class VideoDetailsResponse {
     private Integer file_size;
     private Boolean video_status;
     private String created_at;
-    private Integer streaming_video_id;
-    private Integer cctv_id;
+    private Long streaming_video_id;
+    private Long cctv_id;
     private String cctv_name;
-    private Integer user_id;
+    private Long user_id;
     private String anomaly_type;
 }
